samples/adk_lab/requirements_cloudrun.txt
a2a-sdk
langchain
langgraph
langchain-core
uvicorn
google-cloud-logging
dotenv
google-adk
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-cloud-aiplatform
google-cloud-discoveryengine
google-cloud-logging
google-cloud-secret-manager
google-generativeai
langchain
langchain-community
langchain-core
langgraph
langserve
Pillow
stackapi
uvicorn
